API Funcional v1

+++ database/water_system_schema.sql
@@ -5,7 +5,7 @@
 -- 1. Tabela USUARIOS
 -- Armazena as informações dos usuários que acessarão o sistema.
 CREATE TABLE USUARIOS (
-    id SERIAL PRIMARY KEY,
+    id BIGSERIAL PRIMARY KEY,
     nome VARCHAR(100) NOT NULL,
     email VARCHAR(100) UNIQUE NOT NULL,
     senha_hash VARCHAR(255) NOT NULL, -- Armazenar o hash da senha
@@ -16,15 +16,15 @@ CREATE TABLE USUARIOS (
 -- 2. Tabela PERFIS
 -- Define os níveis de acesso (Ex: Administrador, Usuário Padrão).
 CREATE TABLE PERFIS (
-    id SERIAL PRIMARY KEY,
+    id BIGSERIAL PRIMARY KEY,
     nome_perfil VARCHAR(50) UNIQUE NOT NULL
 );
 
 -- 3. Tabela USUARIO_PERFIS (Tabela de Ligação N:N)
 -- Associa um usuário a um ou mais perfis de permissão.
 CREATE TABLE USUARIO_PERFIS (
-    usuario_id INTEGER NOT NULL,
-    perfil_id INTEGER NOT NULL,
+    usuario_id BIGINT NOT NULL,
+    perfil_id BIGINT NOT NULL,
     PRIMARY KEY (usuario_id, perfil_id),
     FOREIGN KEY (usuario_id) REFERENCES USUARIOS(id) ON DELETE CASCADE,
     FOREIGN KEY (perfil_id) REFERENCES PERFIS(id) ON DELETE CASCADE
@@ -33,8 +33,8 @@ CREATE TABLE USUARIO_PERFIS (
 -- 4. Tabela LOCAIS
 -- Representa os locais onde o consumo de água é medido (ex: Casa, Escritório).
 CREATE TABLE LOCAIS (
-    id SERIAL PRIMARY KEY,
-    usuario_id INTEGER NOT NULL,
+    id BIGSERIAL PRIMARY KEY,
+    usuario_id BIGINT NOT NULL,
     nome_local VARCHAR(100) NOT NULL,
     endereco VARCHAR(255),
     cep VARCHAR(10),
@@ -44,8 +44,8 @@ CREATE TABLE LOCAIS (
 -- 5. Tabela MEDIDORES
 -- Identifica cada hidrômetro (medidor de água) instalado em um local.
 CREATE TABLE MEDIDORES (
-    id SERIAL PRIMARY KEY,
-    local_id INTEGER NOT NULL,
+    id BIGSERIAL PRIMARY KEY,
+    local_id BIGINT NOT NULL,
     codigo_medidor VARCHAR(50) UNIQUE NOT NULL,
     modelo VARCHAR(50),
     data_instalacao DATE,
@@ -55,8 +55,8 @@ CREATE TABLE MEDIDORES (
 -- 6. Tabela LEITURAS
 -- Registra cada leitura de consumo feita no medidor.
 CREATE TABLE LEITURAS (
-    id SERIAL PRIMARY KEY,
-    medidor_id INTEGER NOT NULL,
+    id BIGSERIAL PRIMARY KEY,
+    medidor_id BIGINT NOT NULL,
     data_leitura TIMESTAMP WITH TIME ZONE DEFAULT CURRENT_TIMESTAMP,
     valor_leitura NUMERIC(10, 3) NOT NULL, -- Valor absoluto (Ex: 1234.567 m³)
     foto_leitura VARCHAR(255), -- Caminho/URL da foto da leitura (opcional)
@@ -66,8 +66,8 @@ CREATE TABLE LEITURAS (
 -- 7. Tabela ALERTAS
 -- Guarda os alertas gerados (ex: consumo elevado, suspeita de vazamento).
 CREATE TABLE ALERTAS (
-    id SERIAL PRIMARY KEY,
-    medidor_id INTEGER NOT NULL,
+    id BIGSERIAL PRIMARY KEY,
+    medidor_id BIGINT NOT NULL,
     tipo_alerta VARCHAR(50) NOT NULL, -- Ex: 'CONSUMO_ALTO', 'VAZAMENTO', 'FALHA_LEITURA'
     descricao TEXT,
     data_alerta TIMESTAMP WITH TIME ZONE DEFAULT CURRENT_TIMESTAMP,
